With the collections deriving from ListBase, we no longer need to implement a number of boiler-plate methods.
There are some gotchas with Element and Node lists as they may mutate the list being passed in if it's a nodelist as well, so I'm keeping those unimplemented.

In some cases this may fall back to less efficient operations (such as the WrappedList which no longer forwards every method), but I tried to keep all of the primary cases fast and I believe that these methods would fall back to generic implementations anyways in all cases where we use the classes.

BUG=

Review URL: https://codereview.chromium.org//13852008

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge/dart@21848 260f80e4-7a28-3924-810f-c04153c831b5