Speculative fix for android build.
It is really hard to follow the directory logic, with DART_DIR being just ../../.. and a few pushd and cd commands thrown in.

I don't particularly like this solution either, this is basically to verify that I am assuming correctly in that this is caused by the relative paths being messed up.

Review URL: https://codereview.chromium.org//14961005

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge/dart@22494 260f80e4-7a28-3924-810f-c04153c831b5